moved some functions to helpers folder
